Save file in Delphi before opening it in Vim
If we make a change in Delphi and forget to save, opening the file in
Vim will not have the latest edits in them. This might cause confusion
and conflicts when switching back from Vim to Delphi.

Saving the Delphi file assures that the code in Vim is the same as the
code in Delphi was before switching.